requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-db==0.0.7
idna==3.5
Pillow==10.1.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
autopep8==1.6.0
dj-database-url==0.5.0
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
toml==0.10.2
whitenoise==5.3.0
django-environ==0.8.1
python-dotenv
django-admin-soft-dashboard==1.0.11